Dynamic Context Menu
Define the nodes that will have a context menu and which actions are permitted for each node by defining the 'context' option in the editor's options. For example the options below define that only node1, node2, and all the children of node3 and node4 will have a context menu. Moreover, node1 and node2 allow only the 'Remove' action while the children of node3 and node4 allow both the 'Duplicate' and 'Remove' actions.

context: {
  items: {
    node1: ['Remove'],
    node2: ['Remove']
  },
  children: {
    node3: ['Duplicate', 'Remove'],
    node4: ['Duplicate', 'Remove']
  }
}