Add support for external validation
In some usecases that need control on when a validation is triggered
there is no way to render the resutls of that validation and custom
validation handler cannot be used as that will be called for every
change. E.g., Save/Cancel based validation outside the editor.

Add two new APIs (set/clear)ValidatonErrors to update the results of the
external validation, these errors will be merged with other validation
errors and rendered all together.

Add an example to demonstrate the usecase and document the APIs.