fix: verify object prototype before define 'remove' property in it
Before define 'remove' property in the item's prototype, we need to check whether item is undefined or not.
If we don't check item, it could lead to an error like:
``Cannot use 'in' operator to search for 'remove' in undefined ``
For example, window.DocumentType.prototype can be undefined in some rare environment.